update 진행 중... 골프스타 도와주러...
update 진행 중... 골프스타 도와주러...

+++ APIServer/Model/Table/TableUserGameInfo.cs
@@ -7,8 +7,8 @@ namespace ApiServer.Model
     {
         [StringLength(45)]
         public string ID { get; set; } = "";
-        public Int64 UserLevel { get; set; }
-        public Int64 UserExp { get; set; }
+        public Int32 UserLevel { get; set; }
+        public Int32 UserExp { get; set; }
         public Int32 StarPoint { get; set; } // 별의 모래
     }
 }
